fix: Deprecation UIKit Implementations
