+++ Sources/Montage/3 Component/4 Contents/Card/NormalCard.swift
@@ -247,17 +247,17 @@ import Pretendard
         },
         topContent: {
             HStack {
-                Badge.ContentBadgeController(text: "텍스트")
-                Badge.ContentBadgeController(text: "텍스트")
-                Badge.ContentBadgeController(text: "텍스트")
-                Badge.ContentBadgeController(text: "텍스트")
+                Badge.Content(text: "텍스트")
+                Badge.Content(text: "텍스트")
+                Badge.Content(text: "텍스트")
+                Badge.Content(text: "텍스트")
             }
         },
         bottomContent: {
             HStack {
-                Badge.ContentBadgeController(text: "텍스트")
-                Badge.ContentBadgeController(text: "텍스트")
-                Badge.ContentBadgeController(text: "텍스트")
+                Badge.Content(text: "텍스트")
+                Badge.Content(text: "텍스트")
+                Badge.Content(text: "텍스트")
             }
         }
     )
